--- src/aioway/hop/hop.py
@@ -1,4 +1,0 @@
-# Copyright (c) AIoWay Authors - All Rights Reserved
-
-
-__all__ = []
