+++ src/aioway/op/__init__.py
@@ -1,4 +1,4 @@
 # Copyright (c) AIoWay Authors - All Rights Reserved
 
-from .hop import *
 from .nn import *
+from .op import *
