+++ src/aioway/op/nn/sliding.py
