+++ src/aioway/op/op.py
@@ -1,0 +1,10 @@
+# Copyright (c) AIoWay Authors - All Rights Reserved
+
+__all__ = []
+
+
+class Op:
+    """
+    `Op` stands for [op]erator, or [o]peration [p]review.
+    It is essentailly an unevaluated expression that supports inspection.
+    """
